remove height from SVGELement usage, preload SVGs
