+++ www/src/components/piping/PipingComponent.tsx
@@ -1,6 +1,6 @@
 import { PipingComponentProps } from "../../types/diagram/Piping.ts";
 import { useContext } from "react";
-import PandidContext from "../../context/PandidContext.ts";
+import PandidContext from "../../context/PandidContext.tsx";
 import SvgElement from "../SvgElement.tsx";
 import constructPath from "../../utils/Path.ts";
 import StyledPath from "../StyledPath.tsx";
@@ -17,7 +17,6 @@ export default function PipingComponent(props: PipingComponentProps) {
           id={props.ID}
           componentName={componentName}
           position={props.Position}
-          height={height}
           text={props.GenericAttributes}
         />
       )}
@@ -27,7 +26,6 @@ export default function PipingComponent(props: PipingComponentProps) {
             id={label.ID}
             componentName={label.ComponentName}
             position={label.Position}
-            height={height}
           />
           {label.PolyLine && (
             <StyledPath
