--- www/src/context/PandidContext.ts
@@ -1,10 +1,0 @@
-import { createContext } from "react";
-
-interface PandidContextProps {
-  height: number;
-}
-
-const PandidContext = createContext<PandidContextProps>({
-  height: 0,
-});
-export default PandidContext;
