+++ www/src/hooks/useSerializeNodeSvg.tsx
@@ -1,5 +1,4 @@
-import { useEffect, useState } from "react";
-import useNoakaDexpiSvg from "./useNoakaDexpiSvg.ts";
+import { useContext, useEffect, useState } from "react";
 import {
   addTextToNode,
   removeConnectionPointsAndOrigo,
@@ -7,13 +6,14 @@ import {
   serializeElement,
 } from "../utils/SvgEdit.ts";
 import { GenericAttributesProps } from "../types/diagram/Common.ts";
+import PandidContext from "../context/PandidContext.tsx";
 
 export default function useSerializeNodeSvg(
   componentName: string,
   genericAttributes: GenericAttributesProps | undefined,
 ) {
   const [serializedSvg, setSerializedSvg] = useState<string>("");
-  const svg = useNoakaDexpiSvg(componentName);
+  const svg = useContext(PandidContext).svgMap.get(componentName);
   useEffect(() => {
     if (!svg) return;
     if (genericAttributes) {
